File operation (except read)

+++ IcfsConnector/src/main/java/com/main/DeleteDirectory.java
@@ -25,7 +25,7 @@ public class DeleteDirectory
 
 		// fetch split files and fstats to delete		
 		Map<Integer, FileData> map = FileOperationsDAO.getFilesToDelete(args);
-		
+
 		//test
 		System.out.println("Printing delete list");
 		for(Map.Entry<Integer, FileData> entry : map.entrySet())
@@ -33,35 +33,38 @@ public class DeleteDirectory
 			System.out.println(entry.getKey());
 			System.out.println(entry.getValue());
 		}
-		
+
+		//System.exit(0);
+
 		if(map != null && map.isEmpty())
 		{
 			// directory not found
 			return new ResponseData(0, "Directory not found");
 		}
-		
+
 		// delete from cloud
 		DropBoxIntegration dbi = new DropBoxIntegration();
-		
+
 		for(Map.Entry<Integer, FileData> entry : map.entrySet())
 		{
-			dbi.deleteFilesFromDrive(entry.getValue());
+			if(entry.getValue() != null && !entry.getValue().isDirectory())
+				dbi.deleteFilesFromDrive(entry.getValue());
 		}
-		
+
 		// delete from db and update free space
-		
+
 		if(FileOperationsDAO.deleteFilesFromDB(map))
 		{
 			return new ResponseData(1, "Failed to delete data from db");
-		}		
-		
+		}
+
 		List<DeviceData> devices = FileOperationsDAO.getDeviceData(true);
-		
+
 		if(new FileOperationUtils().updateFreeSpace(devices))
 		{
 			return new ResponseData(1, "Failed to update cloud metadata");
 		}
-		
+
 		return new ResponseData(0, "Delete successful");
 	}
